terra-core: drop dead prefix scaffolding and silence clippy
Prefix cleanup:
- Remove the `prefix_key!` macro and its crate-level re-export. The
  macro was unused outside of its own tests; all production prefix
  types come from `versioned_key!`.
- Drop the default `encode_upper_bound` body from `KeyPrefix` and
  make the method abstract. Every real impl (KeyBound, versioned_key
  output) already overrides it; nothing inherited the default.
- Remove macro-only tests (size / encode_slug_is_hash_only /
  encode_deterministic) and dead-default tests
  (upper_bound_pads_to_key_size / full_prefix_upper_bound_equals_encode).
  Rewrite the lower-bound test with a minimal inline KeyPrefix stub.
- Gate `{Name}Prefix::new()` in `versioned_key!` behind
  `#[cfg(test)] #[allow(dead_code)]` — convenience constructor only
  used by tests, and only by some of them.

Clippy fixes (all terra-core):
- `#[derive(Default)]` on `GetBranchQuery`.
- `impl Default for KeyBound<K>`.
- `Slug::is_empty()` companion to `len()`.
- Drop redundant `u32 -> u32` cast in `time_from_uuid`.
- `repeat_n(0x00u8, pad)` in place of `repeat(0x00u8).take(pad)`.

Workspace builds, tests, fmt --check, and
`clippy --workspace --all-targets -- -D warnings` are all clean.

+++ crates/terra-core/src/domain/tx_meta.rs
@@ -22,5 +22,5 @@ pub struct TxMeta {
 pub fn time_from_uuid(uuid: Uuid) -> Option<DateTime<Utc>> {
     let ts = uuid.get_timestamp()?;
     let (secs, nanos) = ts.to_unix();
-    DateTime::from_timestamp(secs as i64, nanos as u32)
+    DateTime::from_timestamp(secs as i64, nanos)
 }
